+++ public/admin/_app/immutable/nodes/24.CUaYSGPt.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/CxDge0ti.js";export{e as component};